REFACTOR: Fixed the routes and some files for organizations
